requirements.txt
pandas
getopt
email.message
ssl
smtplib
python-dotenv
pathlib
six